requirements.txt
aiofiles==0.8.0
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
multidict==6.0.2
pydantic==1.9.1
python-dotenv==0.20.0
pytz==2022.1
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
typing_extensions==4.3.0
yarl==1.7.2
